+++ test/correctness/PySys_internal_073/run.py
@@ -22,6 +22,9 @@ class PySysTest(BaseTest):
 		runPySys(self, 'sort-by-title', ['print', '--sort', 'title'], workingDir=testsdir) # just check for no exceptions, no validation
 		runPySys(self, 'thistest', ['print', 'PySys_internal_073'], workingDir=testsdir)
 		runPySys(self, 'full', ['print', '--full'], workingDir=testsdir)
+		runPySys(self, 'printTitle', ['print', '--title'], workingDir=testsdir)
+		runPySys(self, 'printDir', ['print', '--dir'], workingDir=testsdir)
+		runPySys(self, 'printTitleAndDir', ['print', '-TD'], workingDir=testsdir)
 		runPySys(self, 'groups', ['print', '--groups'], workingDir=testsdir)
 		runPySys(self, 'modes', ['print', '--modes'], workingDir=testsdir)
 		runPySys(self, 'requirements', ['print', '--requirements'], workingDir=testsdir)
@@ -30,14 +33,24 @@ class PySysTest(BaseTest):
 		runPySys(self, 'emptydir', ['print'], workingDir=self.mkdir('emptydir'), ignoreExitStatus=True, defaultproject=True)
 			
 	def validate(self):
-		for t in ['basic', 'thistest', 'full', 'groups', 'modes']:
+		for t in ['basic', 'thistest', 'full', 'groups', 'modes', 'printTitle', 'printDir', 'printTitleAndDir']:
 			self.assertGrep(t+'.err', expr='.*', contains=False) # no errors
 
 		self.assertGrep('basic.out', expr='PySys_internal_073 *[|] *[^ ]+')
 		self.assertGrep('full.out', expr='Test id *: *PySys_internal_073') # just pick one example
 		self.assertGrep('full.out', expr=r'Test directory *: *correctness[/\\]PySys_internal_073')
 		self.assertGrep('full.out', expr=r"Test user data *: *myUserDataKey='12345', myUserDataKey2='Hello \\\\n world'") # from this test's descriptor
-	
+
+		# check the first line is as expected
+		self.assertGrepOfGrep('printTitle.out',        '.+', '.+ | .+')
+		self.assertGrepOfGrep('printTitleAndDir.out',  '.+', '.+ | .+')
+		self.assertGrep('printDir.out', ' | ', contains=False)
+
+		self.assertGrepOfGrep('printDir.out',          '.+', '^[^ ].*'+re.escape(os.sep))
+		self.assertGrepOfGrep('printTitleAndDir.out',  '.+', '[^ ]+ | .+')
+		self.assertGrep('printTitleAndDir.out',  '  +.+'+re.escape(os.sep))
+
+
 		self.assertGrep('modes.out', expr='MyMode') # just pick one example
 		self.assertLineCount('thistest.out', expr='.', condition='==1')